202011191708代码备份

+++ pages/three/three.js
@@ -1,41 +1,38 @@
 // pages/Three/Three.js
 import { getCollection, toProjectDetail } from '../../api/smallProgram'
-let innerAudioContext = {};
+let innerAudioContext;
+let timer;
 Page({
   data: {
-    // innerAudioContext: {},
-    playStatus: false,
+    playStatus: true,
     src:''
   },
   onLoad: async function (options) {
     const { recno, name } = options
+    const that = this;
     innerAudioContext = wx.createInnerAudioContext()
     wx.setNavigationBarTitle({ title: name })
     const url = `${App.baseApi.Prod}SmallProgram/toCollection3D.do?recNo=${recno}`
     this.setData({ src: url });
     const collection = await getCollection({ recNo: recno })
     this.setData({collection})
-    let that = this
-    setTimeout(function(){
+    timer = setTimeout(function(){
       innerAudioContext.src = collection.mp3Url
-      that.data.playStatus ? innerAudioContext.pause() : innerAudioContext.play()
-      that.setData({ playStatus: !that.data.playStatus })
+      innerAudioContext.play()
     }, 2500)
     innerAudioContext.onEnded(function(){
-      this.setData({ playStatus: !this.data.playStatus })
-      innerAudioContext.stop()
-      innerAudioContext.destroy()
+      that._leaveAndStopMp3();
     })
   },
-
   onUnload() {
     this._leaveAndStopMp3()
   },
   /** 离开页面停止音乐播放 */
   _leaveAndStopMp3() {
-    if (this.data.playStatus) {
-      innerAudioContext.stop()
-      innerAudioContext.destroy()
+    if(innerAudioContext){
+      clearTimeout(timer);
+      innerAudioContext.stop();
+      innerAudioContext.destroy();
     }
   },
